fix(frontend): SearchFilterSchema を client モジュールから切り出し絞り込みが無効化される問題を修正
症状: production の /jobs?companyName=X が絞り込まれず全件返ってくる。

原因調査で判明したこと:
- API 単体は正しく絞り込む
- Next.js は URL の searchParams を認識している
- しかし page.tsx のサーバー側で filter が空のまま渡っている
- page.tsx は SearchFilterSchema.fields を spread して SearchParams スキーマを組み立てているが、
  これが build 後の server 側で空オブジェクトとして評価される挙動
- SearchFilterSchema は "use client" 宣言された JobSearchFilter.tsx から export されていた
- "use client" モジュールから import して spread しない形に変えると直る

修正: スキーマを JobSearchFilter.schema.ts ("use client" なし) に切り出し、
page.tsx はそこから直接 import する。pnpm build && pnpm start で絞り込み復活を確認。
